+++ src/stores/dolls/dolls.ts
@@ -1,10 +1,10 @@
 import { AbstractMesh, AnimationGroup, Bone, InstantiatedEntries, Material, Skeleton } from "@babylonjs/core";
 import { forEach } from "chootils/dist/loops";
 import { defaultPosition as defaultPosition2d } from "chootils/dist/points2d";
+import { Point3D } from "chootils/dist/points3d";
 import { mover3dRefs, mover3dState, moverRefs, moverState } from "repond-movers";
 import {
   AnimationNameByModel,
-  AnyAnimationName,
   AnySpotName,
   BoneNameByModel,
   DollName,
@@ -21,8 +21,13 @@ const HIDDEN_POSITION = { x: 0, y: 0, z: -1000 };
 export default function dolls(prendyAssets: PrendyAssets) {
   const { modelNames, dollNames, modelInfoByName, dollOptions } = prendyAssets;
 
-  const { defaultInRange, makeModelAnimWeightsMoverState, modelMoverRefs, modelOtherMeshesRefs } =
-    get_dollStoreUtils(prendyAssets);
+  const {
+    defaultInRange,
+    makeModelAnimWeightsMoverState,
+    makeToggledMeshesState,
+    modelMoverRefs,
+    modelOtherMeshesRefs,
+  } = get_dollStoreUtils(prendyAssets);
 
   const defaultModelName = modelNames[0];
 
@@ -37,7 +42,8 @@ export default function dolls(prendyAssets: PrendyAssets) {
     return {
       modelName: safeModelName as NonNullable<T_ModelName>, // to reference in refs aswell
       //  New room
-      goalSpotName: null as null | AnySpotName, // when going to new place, start at this spot
+      goalSpotNameAtNewPlace: null as null | AnySpotName, // when going to new place, start at this spot
+      goalPositionAtNewPlace: null as null | Point3D, // when going to new place, start at this spot
       //  Movers
       ...mover3dState("position", {
         value: HIDDEN_POSITION,
@@ -48,25 +54,16 @@ export default function dolls(prendyAssets: PrendyAssets) {
       positionOnScreen: defaultPosition2d(),
       // nowAnimation: animationNames[0] as AnimationNameByModel[T_ModelName],
       // animation Weights mover
-      ...makeModelAnimWeightsMoverState(safeModelName)("animWeights"),
-      nowAnimation: animationNames[0] as AnyAnimationName, // NOTE AnimationNameFromDoll might work here?
+      ...makeModelAnimWeightsMoverState<T_ModelName>(safeModelName as T_ModelName)("animWeights"),
+      toggledMeshes: makeToggledMeshesState<T_ModelName>(safeModelName as T_ModelName),
+      nowAnimation: animationNames[0] as AnimationNameByModel[T_ModelName], // NOTE AnimationNameFromDoll might work here?
       animationLoops: true,
       //
       inRange: defaultInRange(),
+      //
+      isVisible: true,
     };
   };
-
-  // hacky way to get return type from generic function
-  // from Colin at https://stackoverflow.com/questions/50321419/typescript-returntype-of-generic-function
-  class StateReturnType_Generic_Helper<T_A extends DollName, T_B extends ModelName> {
-    // wrapped has no explicit return type so we can infer it
-    wrapped(a: T_A, b: T_B) {
-      return state<T_A, T_B>(a, b);
-    }
-  }
-  type StateReturnType<T_A extends DollName, T_B extends ModelName> = ReturnType<
-    StateReturnType_Generic_Helper<T_A, T_B>["wrapped"]
-  >;
 
   type DollAssetRefs<T_ModelName extends ModelName> = {
     meshes: Record<MeshNameByModel[T_ModelName] | "__root__", AbstractMesh>;
@@ -78,7 +75,7 @@ export default function dolls(prendyAssets: PrendyAssets) {
 
   const refs = <T_DollName extends DollName, T_ModelName extends ModelName>(
     dollName: T_DollName,
-    itemState: StateReturnType<T_DollName, T_ModelName>
+    itemState: ReturnType<typeof state<T_DollName, DollOptions[T_DollName]["model"]>>
   ) => {
     const modelName = itemState.modelName;
 
@@ -90,7 +87,7 @@ export default function dolls(prendyAssets: PrendyAssets) {
       aniGroupsRef: null as null | Record<AnimationNameByModel[T_ModelName], AnimationGroup>,
       assetRefs: null as null | DollAssetRefs<T_ModelName>,
       groundRef: null as null | AbstractMesh,
-      checkCollisions: true, // prevents custom collision checking, like isTouching , and moveWithCollisions he the position state changes, NOTE might be okay to use meshRef.checkCollisions instead
+      canGoThroughWalls: false,
       ...mover3dRefs("position", {
         mass: 20,
         damping: 0.5,
@@ -115,7 +112,7 @@ export default function dolls(prendyAssets: PrendyAssets) {
   // };
 
   type DollStartStates = {
-    [K_DollName in DollName]: StateReturnType<K_DollName, DollOptions[K_DollName]["model"]>;
+    [K_DollName in DollName]: ReturnType<typeof state<K_DollName, DollOptions[K_DollName]["model"]>>;
   };
 
   function makeAutmaticModelDollStartStates() {